perf: split core into num/big/inf subdirs with monomorphic utils
Reorganize src/core/ into three type-domain subdirectories: num (finite
number), big (bigint), and inf (number with +/-Infinity), each holding
its own munkres.ts / munkresB.ts. The dispatcher and shared step5/step5B
stay at the core root.

Give each domain its own monomorphic getMin/partitionByMin: core/num/
utils.ts (number, also imported by inf) and core/big/utils.ts (bigint),
replacing the generic getMin (arrayLike.ts) and partitionByMin
(mutableArrayLike.ts) that were shared across type domains.

Why: when one process solves both number and bigint matrices, the shared
generic utils' comparison sites go polymorphic in V8 and the number path
slows ~1.7-1.85x (measured: 1024x1024 number went 23ms -> 43ms once the
utils were touched by bigint). Per-domain copies keep every call site
monomorphic, recovering the loss. Same rationale as the PR #119 core
split. No behavior change; the string getMin overload (unused in
production) was dropped.
