+++ src/core/big/munkres.ts
@@ -1,14 +1,18 @@
-import type { MatrixLike } from "../types/matrixLike.ts";
-import type { Matching } from "../types/matching.ts";
-import type { MutableArrayLike } from "../types/mutableArrayLike.ts";
+import type { MatrixLike } from "../../types/matrixLike.ts";
+import type { Matching } from "../../types/matching.ts";
+import type { MutableArrayLike } from "../../types/mutableArrayLike.ts";
 
-import { getMin } from "../utils/arrayLike.ts";
-import { partitionByMin } from "../utils/mutableArrayLike.ts";
+import { isBigInt } from "../../utils/is.ts";
+import { getMin, partitionByMin } from "./utils.ts";
 
-import { step4B } from "./numMunkresB.ts";
-import { step5 } from "./shared.ts";
+import { step4B } from "./munkresB.ts";
+import { step5 } from "../shared.ts";
 
-export function exec(matrix: MatrixLike<number>): Matching<number> {
+export function exec(matrix: MatrixLike<number>): Matching<number>;
+export function exec(matrix: MatrixLike<bigint>): Matching<bigint>;
+export function exec<T extends number | bigint>(
+  matrix: MatrixLike<T>,
+): Matching<T> {
   // Get dimensions
   const Y = matrix.length;
   const X = matrix[0]?.length ?? 0;
@@ -19,19 +23,23 @@ export function exec(matrix: MatrixLike<number>): Matching<number> {
   }
 
   // Step 1: Reduce
-  const dualX = new Array<number>(X);
-  const dualY = new Array<number>(Y);
+  const dualX = new Array<T>(X);
+  const dualY = new Array<T>(Y);
+  // @ts-expect-error ts(2769)
   step1(matrix, dualX, dualY);
 
   // Steps 2 & 3: Find initial matching
   const starsX = new Array<number>(X).fill(-1);
   const starsY = new Array<number>(Y).fill(-1);
+  // @ts-expect-error ts(2769)
   const stars = steps2To3(matrix, dualX, dualY, starsX, starsY);
 
   // Step 4: Find complete matching
   if (Y <= X) {
+    // @ts-expect-error ts(2769)
     step4(Y - stars, matrix, dualX, dualY, starsX, starsY);
   } else {
+    // @ts-expect-error ts(2769)
     step4B(X - stars, matrix, dualX, dualY, starsX, starsY);
   }
 
@@ -53,25 +61,36 @@ export function exec(matrix: MatrixLike<number>): Matching<number> {
  */
 export function step1(
   matrix: MatrixLike<number>,
-  dualX: number[],
-  dualY: number[],
+  dualX: MutableArrayLike<number>,
+  dualY: MutableArrayLike<number>,
+): void;
+export function step1(
+  matrix: MatrixLike<bigint>,
+  dualX: MutableArrayLike<bigint>,
+  dualY: MutableArrayLike<bigint>,
+): void;
+export function step1<T extends number | bigint>(
+  matrix: MatrixLike<T>,
+  dualX: MutableArrayLike<T>,
+  dualY: MutableArrayLike<T>,
 ): void {
   const X = dualX.length;
   const Y = dualY.length;
 
   // If matrix is tall, skip row reduction
   if (Y > X) {
-    dualY.fill(0);
+    dualY.fill((isBigInt(matrix[0][0]) ? 0n : 0) as T);
   } else {
     // Reduce rows
     for (let y = 0; y < Y; ++y) {
+      // @ts-expect-error ts(2769)
       dualY[y] = getMin(matrix[y])!;
     }
   }
 
   // If matrix is wide, skip column reduction
   if (Y < X) {
-    dualX.fill(0);
+    dualX.fill((isBigInt(matrix[0][0]) ? 0n : 0) as T);
     return;
   }
 
@@ -79,7 +98,7 @@ export function step1(
   let dy = dualY[0];
   let row = matrix[0];
   for (let x = 0; x < X; ++x) {
-    dualX[x] = row[x] - dy || 0;
+    dualX[x] = (row[x] - dy) as T;
   }
 
   // Reduce columns
@@ -87,9 +106,9 @@ export function step1(
     dy = dualY[y];
     row = matrix[y];
     for (let x = 0; x < X; ++x) {
-      const dx = row[x] - dy || 0;
+      const dx = row[x] - dy;
       if (dx < dualX[x]) {
-        dualX[x] = dx;
+        dualX[x] = dx as T;
       }
     }
   }
@@ -108,10 +127,24 @@ export function step1(
  */
 export function steps2To3(
   matrix: MatrixLike<number>,
-  dualX: number[],
-  dualY: number[],
-  starsX: number[],
-  starsY: number[],
+  dualX: ArrayLike<number>,
+  dualY: ArrayLike<number>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
+): number;
+export function steps2To3(
+  matrix: MatrixLike<bigint>,
+  dualX: ArrayLike<bigint>,
+  dualY: ArrayLike<bigint>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
+): number;
+export function steps2To3<T extends number | bigint>(
+  matrix: MatrixLike<T>,
+  dualX: ArrayLike<T>,
+  dualY: ArrayLike<T>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
 ): number {
   const X = dualX.length;
   const Y = dualY.length;
@@ -122,7 +155,7 @@ export function steps2To3(
     const dy = dualY[y];
     const row = matrix[y];
     for (let x = 0; x < X; ++x) {
-      if (starsX[x] === -1 && row[x] === (dualX[x] + dy || 0)) {
+      if (starsX[x] === -1 && dy === row[x] - dualX[x]) {
         starsX[x] = y;
         starsY[y] = x;
         ++stars;
@@ -149,10 +182,26 @@ export function steps2To3(
 export function step4(
   unmatched: number,
   matrix: MatrixLike<number>,
-  dualX: number[],
-  dualY: number[],
-  starsX: number[],
-  starsY: number[],
+  dualX: MutableArrayLike<number>,
+  dualY: MutableArrayLike<number>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
+): void;
+export function step4(
+  unmatched: number,
+  matrix: MatrixLike<bigint>,
+  dualX: MutableArrayLike<bigint>,
+  dualY: MutableArrayLike<bigint>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
+): void;
+export function step4<T extends number | bigint>(
+  unmatched: number,
+  matrix: MatrixLike<T>,
+  dualX: MutableArrayLike<T>,
+  dualY: MutableArrayLike<T>,
+  starsX: MutableArrayLike<number>,
+  starsY: MutableArrayLike<number>,
 ): void {
   // If no unmatched row
   if (unmatched <= 0) {
@@ -161,7 +210,7 @@ export function step4(
 
   const X = dualX.length;
   const slack = new Uint32Array(X);
-  const slackV = new Array<number>(X);
+  const slackV = new Array<T>(X);
   const slackY = new Uint32Array(X);
 
   // Match unmatched rows
@@ -170,10 +219,12 @@ export function step4(
       continue;
     }
 
+    // @ts-expect-error ts(2769)
     const N = match(y, matrix, dualX, dualY, starsX, slack, slackV, slackY);
     --unmatched;
 
     // Update dual variables
+    // @ts-expect-error ts(2769)
     step6(y, N, dualX, dualY, slack, slackV, starsX);
 
     // Update matching
@@ -195,21 +246,42 @@ export function step4(
 export function step6(
   y: number,
   N: number,
-  dualX: number[],
-  dualY: number[],
+  dualX: MutableArrayLike<number>,
+  dualY: MutableArrayLike<number>,
   slack: ArrayLike<number>,
   slackV: ArrayLike<number>,
   starsX: ArrayLike<number>,
+): void;
+export function step6(
+  y: number,
+  N: number,
+  dualX: MutableArrayLike<bigint>,
+  dualY: MutableArrayLike<bigint>,
+  slack: ArrayLike<number>,
+  slackV: ArrayLike<bigint>,
+  starsX: ArrayLike<number>,
+): void;
+export function step6<T extends number | bigint>(
+  y: number,
+  N: number,
+  dualX: MutableArrayLike<T>,
+  dualY: MutableArrayLike<T>,
+  slack: ArrayLike<number>,
+  slackV: ArrayLike<T>,
+  starsX: ArrayLike<number>,
 ): void {
   const sum = slackV[slack[--N]];
-  dualY[y] = dualY[y] + sum || 0;
+  // @ts-expect-error ts(2365)
+  dualY[y] += sum;
 
   for (let i = 0; i < N; ++i) {
     const x = slack[i];
     y = starsX[x];
-    const min = sum - slackV[x] || 0;
-    dualX[x] = dualX[x] - min || 0;
-    dualY[y] = dualY[y] + min || 0;
+    const min = sum - slackV[x];
+    // @ts-expect-error ts(2322)
+    dualX[x] -= min;
+    // @ts-expect-error ts(2365)
+    dualY[y] += min;
   }
 }
 
@@ -228,11 +300,31 @@ export function step6(
 export function match(
   y: number,
   matrix: MatrixLike<number>,
-  dualX: number[],
-  dualY: number[],
-  starsX: number[],
+  dualX: ArrayLike<number>,
+  dualY: ArrayLike<number>,
+  starsX: ArrayLike<number>,
   slack: MutableArrayLike<number>,
   slackV: MutableArrayLike<number>,
+  slackY: MutableArrayLike<number>,
+): number;
+export function match(
+  y: number,
+  matrix: MatrixLike<bigint>,
+  dualX: ArrayLike<bigint>,
+  dualY: ArrayLike<bigint>,
+  starsX: ArrayLike<number>,
+  slack: MutableArrayLike<number>,
+  slackV: MutableArrayLike<bigint>,
+  slackY: MutableArrayLike<number>,
+): number;
+export function match<T extends number | bigint>(
+  y: number,
+  matrix: MatrixLike<T>,
+  dualX: ArrayLike<T>,
+  dualY: ArrayLike<T>,
+  starsX: ArrayLike<number>,
+  slack: MutableArrayLike<number>,
+  slackV: MutableArrayLike<T>,
   slackY: MutableArrayLike<number>,
 ): number {
   const X = slack.length;
@@ -242,11 +334,12 @@ export function match(
   let row = matrix[y];
   for (let x = 0; x < X; ++x) {
     slack[x] = x;
-    slackV[x] = row[x] - (dualX[x] + dy || 0) || 0;
+    slackV[x] = (row[x] - dualX[x] - dy) as T;
     slackY[x] = y;
   }
 
   // Initialize zeros
+  // @ts-expect-error ts(2769)
   let zeros = partitionByMin(slack, slackV, 0);
   let zero = slackV[slack[0]];
 
@@ -255,11 +348,11 @@ export function match(
   for (let x = slack[0]; starsX[x] !== -1; x = slack[steps++]) {
     // Update slack
     y = starsX[x];
-    dy = dualY[y];
+    dy = (dualY[y] - zero) as T;
     row = matrix[y];
     for (let i = zeros; i < X; ++i) {
       x = slack[i];
-      const value = (row[x] - (dualX[x] + dy || 0) || 0) + zero || 0;
+      const value = (row[x] - dualX[x] - dy) as T;
       if (value >= slackV[x]) {
         continue;
       }
@@ -273,6 +366,7 @@ export function match(
 
     // Update zeros
     if (steps >= zeros) {
+      // @ts-expect-error ts(2769)
       zeros = partitionByMin(slack, slackV, zeros);
       zero = slackV[slack[steps]];
     }
